Refactor : PreviewArticles => PreviewArticle 수정
console.log도 삭제, 쓰이지 않던 함수 삭제
